Move ModelResolutionException to exceptions package
