+++ core/esmf-aspect-meta-model-java/src/main/java/org/eclipse/esmf/aspectmodel/resolver/CommandExecutor.java
@@ -18,6 +18,8 @@ import java.io.InputStream;
 import java.util.Scanner;
 import java.util.StringTokenizer;
 
+import org.eclipse.esmf.aspectmodel.resolver.exceptions.ModelResolutionException;
+
 /**
  * Executes an external resolver via the underlying OS command and returns the stdout from the command as result.
  */
